fixed compiler warning in visual studio
--HG--
extra : convert_revision : svn%3Afdd8eb12-d10e-0410-9acb-85c331704f74/trunk%402643

+++ dlib/threads/thread_pool_extension.cpp
@@ -216,7 +216,7 @@ namespace dlib
         uint64 id
     ) const
     {
-        return id%tasks.size();
+        return static_cast<unsigned long>(id%tasks.size());
     }
 
 // ----------------------------------------------------------------------------------------
